requirements-ml.txt
# Optional: CNN battlefield detector and training script (scripts/train/train_battlefield_classifier.py).
torch>=2.1.0
Pillow>=9.0
numpy>=1.24.0
